wfcatom and swfcatom have two indices also in the noncollinear case, in analogy to evc.
Apparently swfcatom_nc was used but not allocated in noncollinear
calculations with LDA+U (not sure it was implemented, though)

There is still something I don't understand in wfcinit: the starting
wavefunctions needs to be set to 0 for n>npw at the end of the routine,
while it should be sufficient to clean them at the beginning. This might
be a sign that we are overwriting a zone that should not be overwritten.