Fixing an overflow bug:
  for(i=0;i-len-1;i++) strHelper[i]=' '

This was probably a typo and i<len-1 was meant.
This overflow bug had a very severe consequence on one of my
machine: all negative values were parsed as being positive.

Adding a check that a maximum size of len is 256, and also the length
of strHelper/strError strings changed to 257, or the last char in
strExpression was eaten. For more detailes see the file.