Added an optional variable in diropn. If diropn is called in the usual way:
CALL diropn (iudvrot, fildvscfrot, lrdrho, exst)

then tmp_dir is used as directory for the file to be opened,
as usual. 

If tmp_dir_ is specified in the call

CALL diropn (iudvrot, fildvscfrot, lrdrho, exst, tmp_dir_)

then tmp_dir_ is used as directory for the file to be opened.

Matteo Calandra