Bugfix for open_close_input_file.f90 .
Variable input_file was not being initialized so in the case of pw.x being called with no additional arguemtns input being provided via stdin the uninitalized value of input_file was being written to output. The variable is now initalized as a null string and furthermore if stdin is used this is printed in the output.