The introduction of a new variable in dfile_name implies changes in all the place where dfile_name is called: dfile_star io_pattern solve_linter dynmatrix
It is necessary to modify
dynmatrix and phonon.f90 in such a way that an integer variable iq is passed
to the routine: 
call dynmatrix(iq)
In the case ldisp=.true. and electro_phonon='Wannier' iq is
the index of the phonon momenta in the reduced Brillouin zone
(1,...,nqs).

MCB