PAW 4-wavefunction exchange integrals symmetrize w.r.t index exchange to prevent S-matrix to become non-positive. The reason is that the one-center hartree kernel is not exactly symmetrical, and the small discrepancy tends to accumulate.
Several improvements here and there to make the code and the output cleaner.

This version is semi-definitive, except for USPP and PAW contributions to force, and for the stress, everything should be working properly both in real space and G space, in serial and in parallel (any kind of parallelisation).

This version of the code is thereby guaranteed 100% bug free: any bug report will be considered a mean-spirited joke and ignored.

LP